refactor(span-icons): Move next to SpanBarRow and improve tests
Move span-icons.ts and span-icons.test.ts from src/utils/ to
src/components/TracePage/TraceTimelineViewer/ where its only
consumer SpanBarRow.tsx lives.

Rewrite tests to check behavior (non-null for recognized namespaces,
null for unknown/empty, correct priority ordering) rather than pinning
specific icon component identities.
